conversordpi/conversordpi/requirements.txt
gradio
Pillow
numpy

